auth/requirements.txt
fastapi==0.115.11
psycopg==3.2.6
uvicorn==0.34.0
pydantic==2.10.6
asyncio==3.4.3
argon2-cffi==23.1.0
biscuit-python==0.3.2